Move `InvalidMutatorException` from `Infection\Exception` to `Infection\Mutator`
